MNIST posterior fig

+++ sample_condition_ddrm.py
@@ -220,7 +220,7 @@ def main():
                 with torch.no_grad():
                     x_start = torch.randn(ref_img.shape, device=device)
                     sample, return_ims = sample_fn(x_start=x_start, measurement=y_n, record=True, save_root=out_path, mask=mask,
-                                       noise_sig=measure_config['noise']['sigma'], meas_type=measure_config['operator']['name'], truth=ref_img, svd=H).float()
+                                       noise_sig=measure_config['noise']['sigma'], meas_type=measure_config['operator']['name'], truth=ref_img, svd=H)
                     im_count = 0
                     for im in return_ims:
                         plt.imsave(f'motivation_fig/interediate_ddrm_{i}_{im_count}.png', clear_color(im))
